backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.6.1  # Version compatible with langchain-openai
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pgvector==0.2.0
pdfplumber==0.10.3
pypdf==3.17.0
langchain==0.0.340
langchain-openai==0.0.2
numpy==1.24.3
pydantic==2.5.0
pydantic-settings==2.1.0
alembic==1.12.1
chromadb==0.4.22